+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Added `JsonVariant::as<char*>()` as a synonym for `JsonVariant::as<const char*>()` (issue #257)
 * Added example `JsonHttpClient` (issue #256)
+* Added `JsonArray::copyTo()` and `JsonArray::copyFrom()` (issue #254)
 
 v5.1.1
 ------